% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rcpp_hello_world}
\alias{rcpp_hello_world}
\title{rcpp_hello_world.}
\usage{
rcpp_hello_world()
}
\arguments{
\item{x}{A description of the parameter 'x'. The
description can span multiple lines.}

\item{y}{A description of the parameter 'y'.}
}
\description{
A example of a function written in C++.
}
\details{
These are further details.
}
\section{Headers}{


Note header is pure C++, is not exported to RCpp or to NAMESPACE. Is usage is completely internal.
}

